seperated viewModels Long functions from it, by introducing ViewModelsLogic classes
