+++ app/src/main/java/com/example/juzzics/features/home/HomeScreen.kt
@@ -1,5 +1,6 @@
 package com.example.juzzics.features.home
 
+import android.widget.Toast
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,14 +8,14 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import com.example.juzzics.common.base.extensions.with2
 import com.example.juzzics.common.base.viewModel.Action
 import com.example.juzzics.common.base.viewModel.UiEvent
 import com.example.juzzics.common.base.viewModel.ViewState
-import com.example.juzzics.common.base.viewModel.getState
+import com.example.juzzics.common.base.viewModel.listen
 import com.example.juzzics.common.base.viewModel.state
 import kotlinx.coroutines.flow.SharedFlow
 
@@ -25,6 +26,14 @@ fun HomeScreen(
     onAction: (Action) -> Unit
 ) {
     with2(first = states, second = HomeVM) {
+        val context = LocalContext.current
+        uiEvent.listen {
+            when (it) {
+                is HomeVM.ToastMsg -> {
+                    Toast.makeText(context, it.msg, Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
         Surface {
             Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
                 Text(text = TEST.state() ?: "Home Screen", modifier = Modifier.clickable {
